Fix bag
1. Changed the name of class "Dictionary" to class "Console";
2. Added class WorkWithFile
